Removed duplicate tests, moved tests to integration file, removed comments
